More Memory Management Stuff
Added a function to deallocate the `data_list` member of a `server_info` instance. Also, added the `data_list` member data to the server struct to give the object a container.

+++ inc/list.h
@@ -1,13 +1,11 @@
 #ifndef SCF_LIST_H
 #define SCF_LIST_H
 
-#include <stdint.h>
-
 struct list_member
 {
     struct list_member *next, *previous;
     void *data;
-    uint64_t data_size;
+    unsigned int data_size;
 };
 
 struct data_list
